+++ src/ByteDev.Cmd/Arguments/CmdAllowedArg.cs
@@ -1,4 +1,6 @@
-﻿using System.Text;
+﻿using System;
+using System.Text;
+using System.Text.RegularExpressions;
 
 namespace ByteDev.Cmd.Arguments
 {
@@ -9,13 +11,19 @@ namespace ByteDev.Cmd.Arguments
     {
         private const string ArgNamePrefix = "-";
 
+        private string _longName;
+
         /// <summary>
         /// Initializes a new instance of the <see cref="T:ByteDev.Cmd.Arguments.CmdAllowedArg" /> class.
         /// </summary>
         /// <param name="shortName">Argument short name (a single character).</param>
-        /// <param name="hasValue">Whether the argument should have a value.</param>
+        /// <param name="hasValue">Whether the argument is expected to have a value.</param>
+        /// <exception cref="T:System.ArgumentException"><paramref name="shortName" /> is an invalid character.</exception>
         public CmdAllowedArg(char shortName, bool hasValue)
         {
+            if (!char.IsLetter(shortName))
+                throw new ArgumentException("Argument short name must be only a A-Z or a-z character.", nameof(shortName));
+
             ShortName = shortName;
             HasValue = hasValue;
         }
@@ -31,9 +39,20 @@ namespace ByteDev.Cmd.Arguments
         public bool HasValue { get; }
 
         /// <summary>
-        /// Optional long name for the argument.
+        /// Optional long name for the argument. Must contain only characters [A-Za-z] or be set to null.
         /// </summary>
-        public string LongName { get; set; }
+        /// <exception cref="T:System.ArgumentException"><paramref name="value" /> contains invalid characters.</exception>
+        public string LongName
+        {
+            get => _longName;
+            set
+            {
+                if (value != null && !Regex.IsMatch(value, "^[A-Za-z]+$"))
+                    throw new ArgumentException("Argument long name must contain only A-Z or a-z characters.", nameof(value));
+                
+                _longName = value;
+            }
+        }
 
         /// <summary>
         /// Optional description of the argument. Used in help text for the user.